+++ SFML 2.3.21/Box.cpp
@@ -7,7 +7,7 @@ ObjectType Box::getType() const
 Box::Box(const Resources& res, const b2Vec2& size, const b2Vec2& position) :
 	mSize(size),
 	mSprite(res.textures.get(Texture::SPRITE_BOX),
-			sf::IntRect(0, 0, size.x*PIXELS_PER_METER, size.y*PIXELS_PER_METER)
+			sf::IntRect(0, 0, static_cast<int>(size.x*PIXELS_PER_METER), static_cast<int>(size.y*PIXELS_PER_METER))
 	)
 {
 	auto bounds = mSprite.getLocalBounds();
