+++ SFML 2.3.21/DeathBlock.cpp
@@ -30,7 +30,7 @@ void DeathBlock::draw(sf::RenderTarget & target, sf::RenderStates states) const
 DeathBlock::DeathBlock(const Resources& res, const b2Vec2& size, const b2Vec2& position) :
 	mSize(size),
 	mSprite(res.textures.get(Texture::SPRITE_LAVA),
-			sf::IntRect(0, 0, size.x*PIXELS_PER_METER, size.y*PIXELS_PER_METER)
+			sf::IntRect(0, 0, static_cast<int>(size.x*PIXELS_PER_METER), static_cast<int>(size.y*PIXELS_PER_METER))
 	)
 {
 	auto bounds = mSprite.getLocalBounds();
